% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{error}
\alias{error}
\alias{fatal}
\alias{info}
\alias{warn}
\title{Print Messages}
\usage{
error(error_name, msg = NULL)

fatal(error_name, msg = NULL)

info(error_name, msg = NULL)

warn(error_name, msg = NULL)
}
\description{
Print Messages
}
