% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_advanced_function_hist.R
\name{hist.rodps.vector}
\alias{hist.rodps.vector}
\title{create odps.data and odps.vector in S4}
\usage{
\method{hist}{rodps.vector}(ov)
}
\description{
create odps.data and odps.vector in S4
}
