% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_data_structure.R
\name{init.odps.ext}
\alias{init.odps.ext}
\title{create odps.data and odps.vector in S4
tbl: table name, does not accept partition
load.time: when object is created, to determin if the buffer data is outdated
data: local buffer, if data is small enough, load it into buffer}
\usage{
init.odps.ext()
}
\description{
create odps.data and odps.vector in S4
tbl: table name, does not accept partition
load.time: when object is created, to determin if the buffer data is outdated
data: local buffer, if data is small enough, load it into buffer
}
