% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_base_function.R
\name{rodps.bizid}
\alias{rodps.bizid}
\title{Set business ID}
\usage{
rodps.bizid(bizid)
}
\arguments{
\item{bizid}{business id, e.g. 012345^.}
}
\description{
Set business ID
}
\examples{
## set business id to 012345
\dontrun{rodps.bizid('012345^')}
}
\seealso{
\code{\link{RODPS}}, \code{\link{rodps.sql}}
}
\author{
\email{ruibo.lirb@alibaba-inc.com}
}
