% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_project.R
\name{rodps.current.project}
\alias{rodps.current.project}
\title{Get current project name.}
\usage{
rodps.current.project()
}
\description{
Get current project name.
}
\seealso{
\code{\link{rodps.project.current}}
}
