% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.generate.DDL}
\alias{rodps.generate.DDL}
\title{DDL Generation}
\usage{
rodps.generate.DDL(full.tablename, dataframe, tablecomment = NULL)
}
\arguments{
\item{full.tablename}{Table name, in format 'ProjectName.TableName',or
'TableName' (using current project).}

\item{dataframe}{Source data frame.}

\item{tablecomment}{DDL comment string.}
}
\description{
Generate SQL DDL from dataframe.
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
