% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rodps.init}
\alias{rodps.init}
\title{Init ODPS with configs}
\usage{
rodps.init(path = NULL, access.id = NULL, access.key = NULL)
}
\arguments{
\item{path}{File path string indicating odps_config}

\item{access.id}{Access ID string}

\item{access.key}{Access Key string}
}
\description{
Init ODPS with configs
}
