% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_advanced_function_predict.R
\name{rodps.predict.fda}
\alias{rodps.predict.fda}
\title{Predict fda model}
\usage{
\method{rodps.predict}{fda}(object, srctbl, tgttbl, type = "class", prior, dimension = 2)
}
\description{
Predict fda model
}
