% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_predict.R
\name{rodps.predict.rpart}
\alias{rodps.predict.rpart}
\title{Extend Recursive Partitioning}
\usage{
\method{rodps.predict}{rpart}(object, srctbl, tgttbl, inc.col = NULL, dryrun = FALSE)
}
\arguments{
\item{object}{Rpart model}

\item{srctbl}{Data source table}

\item{tgttbl}{Target table of prediction results}

\item{dryrun}{Return the prediction SQL string instead of running the query}
}
\description{
Extend Recursive Partitioning
}
