% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_project.R
\name{rodps.project.current}
\alias{rodps.project.current}
\title{Get current project name.}
\usage{
rodps.project.current()
}
\description{
Get current project name.
}
\examples{
## get current project name
\dontrun{rodps.project.current()}
}
\seealso{
\code{\link{rodps.project.use}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
