% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_project.R
\name{rodps.project}
\alias{rodps.project}
\title{Project functions}
\description{
Provide functions to operate project.
}
\seealso{
\code{\link{rodps.project.use}},
  \code{\link{rodps.project.current}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
