% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_project.R
\name{rodps.project.use}
\alias{rodps.project.use}
\title{Change current project.}
\usage{
rodps.project.use(projectname)
}
\arguments{
\item{projectname}{target projectname; make sure that you have the authority to access this Project.}
}
\description{
Change current project.
}
\examples{
## change project to prjb
\dontrun{rodps.project.use('prjb')}
}
\seealso{
\code{\link{rodps.project.current}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
