% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_base_function.R
\name{rodps.set}
\alias{rodps.set}
\title{Set task properties}
\usage{
rodps.set(key, value)
}
\arguments{
\item{key}{setting name, e.g. odps.sql.allow.fullscan.}

\item{value}{setting value.}
}
\description{
Set properties for SQL task
}
\examples{
## enable full table scan
\dontrun{rodps.set('odps.sql.allow.fullscan', 'true')}
}
\seealso{
\code{\link{RODPS}}, \code{\link{rodps.sql}},
  \code{\link{rodps.unset}},
}
\author{
\email{ruibo.lirb@alibaba-inc.com}
}
