% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.split.ftn}
\alias{rodps.split.ftn}
\title{Split full table name into table name and project name}
\usage{
rodps.split.ftn(ftn)
}
\arguments{
\item{ftn}{Full table name.}
}
\description{
Split full table name into table name and project name
}
