% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_sql.R
\name{rodps.sql}
\alias{rodps.sql}
\alias{rodps.query}
\title{SQL Command}
\usage{
rodps.sql(query, mcqa = FALSE)

rodps.query(query, mcqa = FALSE)
}
\arguments{
\item{query}{SQL string}

\item{mcqa}{Whether enable MCQA or not}
}
\description{
Run SQL command and return result(in data.frame type).
}
\examples{
## select the data of 'sales' in January ,and store the result in data.frame
\dontrun{ data <- rodps.sql('select * from sales where month=1')}
}
\seealso{
\code{\link{RODPS}}, \code{\link{rodps.table}},
  \code{\link{rodps.project}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
