% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_str.R
\name{rodps.str}
\alias{rodps.str}
\title{Display Table}
\usage{
rodps.str(tbl)
}
\arguments{
\item{tbl}{RODPS Table object}
}
\value{
Formatted string.
}
\description{
Print table as formatted string.
}
\seealso{
[str.rodps.data()]
}
