% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.desc}
\alias{rodps.table.desc}
\alias{rodps.desc.table}
\title{Table Description}
\usage{
rodps.table.desc(full.tablename, partition = NULL)

rodps.desc.table(full.tablename, partition = NULL)
}
\arguments{
\item{full.tablename}{Table name, in format 'ProjectName.TableName',or
'TableName' (using current project).}

\item{partition}{Partition spec}
}
\description{
Show description of a table, including metadata of 
Owner, Project, Comment, Create_time, Last_modified_time, Size, Columns.
}
\examples{
## show description of 'dual'
\dontrun{rodps.table.desc('dual')}
}
\seealso{
\code{\link{rodps.table.drop}}, \code{\link{rodps.table.exist}},
  \code{\link{rodps.table.partitions}}, \code{\link{rodps.table.list}},
  \code{\link{rodps.table.rows}}, \code{\link{rodps.table.size}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
