% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.drop}
\alias{rodps.table.drop}
\alias{rodps.drop.table}
\title{Drop Table}
\usage{
rodps.table.drop(full.tablename, partition = NULL)

rodps.drop.table(full.tablename, partition = NULL)
}
\arguments{
\item{full.tablename}{Table name.}

\item{partition}{Partition spec.}
}
\description{
Delete table if it exists.
}
\examples{
\dontrun{rodps.table.drop('sales_backup')}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.exist}},
  \code{\link{rodps.table.partitions}}, \code{\link{rodps.table.list}},
  \code{\link{rodps.table.rows}}, \code{\link{rodps.table.size}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
