% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.exist}
\alias{rodps.table.exist}
\alias{rodps.exist.table}
\title{Table Existence}
\usage{
rodps.table.exist(full.tablename, partition = NULL)

rodps.exist.table(full.tablename, partition = NULL)
}
\arguments{
\item{full.tablename}{table name.}

\item{partition}{partition spec, default NULL.}
}
\description{
Query whether the table exists.
}
\examples{
\dontrun{rodps.table.exist('mytable')}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.partitions}}, \code{\link{rodps.table.list}},
  \code{\link{rodps.table.rows}}, \code{\link{rodps.table.size}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
