% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.head}
\alias{rodps.table.head}
\title{Table Head}
\usage{
rodps.table.head(tbl, n = 6L)
}
\arguments{
\item{tbl}{Table name}

\item{n}{The number of head rows}
}
\description{
Show a few of head rows of table.
}
