% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table_hist.R
\name{rodps.table.hist}
\alias{rodps.table.hist}
\alias{rodps.hist}
\title{Table Histogram}
\usage{
rodps.table.hist(
  tblname,
  colname,
  breaks = NULL,
  freq = TRUE,
  include.lowest = TRUE,
  right = TRUE,
  main = paste("Histogram of ", colname),
  xlab = colname,
  ...
)

rodps.hist(
  tblname,
  colname,
  breaks = NULL,
  freq = TRUE,
  include.lowest = TRUE,
  right = TRUE,
  main = paste("Histogram of ", colname),
  xlab = colname,
  ...
)
}
\description{
Extend hist. This function returns a list of breaks, counts, density, mids, xname,
equidist, class attr from tbl; then plot with the list
}
