% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.list}
\alias{rodps.table.list}
\alias{rodps.list.table}
\alias{rodps.list.tables}
\title{List Tables}
\usage{
rodps.table.list(pattern = NULL, projectname = NULL)

rodps.list.table(pattern = NULL, projectname = NULL)

rodps.list.tables(pattern = NULL, projectname = NULL)
}
\arguments{
\item{pattern}{Partition pattern, use '*' or specific PartitionName.}

\item{projectname}{Specific project to query,default is current project.}
}
\description{
List tables in the project, default in current project.
}
\examples{
##list the tables in current project
\dontrun{rodps.table.list()}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
  \code{\link{rodps.table.rows}}, \code{\link{rodps.table.size}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
