% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table}
\alias{rodps.table}
\title{Table functions}
\description{
Provide functions to operate table.
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
\code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
\code{\link{rodps.table.list}}, \code{\link{rodps.table.rows}},
\code{\link{rodps.table.size}}, \code{\link{rodps.table.read}},
\code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
