% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.read}
\alias{rodps.table.read}
\alias{rodps.read.table}
\alias{rodps.load.table}
\title{Reading Table}
\usage{
rodps.table.read(
  full.tablename,
  partition = NULL,
  limit = -1,
  memsize = 10737518240,
  isdebug = FALSE,
  thread = 8
)

rodps.read.table(
  full.tablename,
  partition = NULL,
  limit = -1,
  memsize = 10737518240,
  isdebug = FALSE,
  thread = 8
)

rodps.load.table(
  full.tablename,
  partition = NULL,
  limit = -1,
  memsize = 10737518240,
  isdebug = FALSE,
  thread = 8
)
}
\arguments{
\item{full.tablename}{Table name}

\item{partition}{Partition spec}

\item{limit}{Limit the rows to read, '-1' for not limit.}

\item{memsize}{Maximum data capacity.}

\item{isdebug}{Boolean value, if debugging is enabled.}

\item{thread}{Thread number.}
}
\description{
Read data from ODPS and store in R data format.
}
\examples{
## show description of 'dual'
\dontrun{ x<-rodps.table.read('sales',partition='ds=20180124',limit=100) }
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
  \code{\link{rodps.table.list}}, \code{\link{rodps.table.rows}},
  \code{\link{rodps.table.size}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
