% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.rows}
\alias{rodps.table.rows}
\alias{rodps.rows.table}
\title{Table Rows}
\usage{
rodps.table.rows(full.tablename, partition = NULL)

rodps.rows.table(full.tablename, partition = NULL)
}
\arguments{
\item{full.tablename}{Table name , in format of 'ProjectName.TableName' or
'TableName' (using current project)}

\item{partition}{Partition spec.}
}
\description{
Get the number of rows in a table.
}
\examples{
## get the number of rows
\dontrun{rodps.table.rows('sales')}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
  \code{\link{rodps.table.list}}, \code{\link{rodps.table.size}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
