% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table_sample.R
\name{rodps.table.sample.srs}
\alias{rodps.table.sample.srs}
\alias{rodps.sample.srs}
\title{Sample table}
\usage{
rodps.table.sample.srs(
  srctable,
  tgttable,
  samplerate,
  cond = NULL,
  select = NULL
)

rodps.sample.srs(srctable, tgttable, samplerate, cond = NULL, select = NULL)
}
\description{
Sample table
}
\seealso{
\code{\link{rodps.table.sample.strat}}
}
