% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.sample.strat}
\alias{rodps.table.sample.strat}
\title{分层抽样 select abc from ( *, row_number() over( partition by g order by
rand()) r_rn, rand() as r_select ) sub
1. by percent sub where r_select < rate
2. by_number sub where rn <= rate}
\usage{
rodps.table.sample.strat(srctable, tgttable, samplerate, strat, select = NULL)
}
\description{
分层抽样 select abc from ( *, row_number() over( partition by g order by
rand()) r_rn, rand() as r_select ) sub
1. by percent sub where r_select < rate
2. by_number sub where rn <= rate
}
