% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.size}
\alias{rodps.table.size}
\alias{rodps.size.table}
\title{Table Size}
\usage{
rodps.table.size(full.tablename, partition = NULL)

rodps.size.table(full.tablename, partition = NULL)
}
\arguments{
\item{full.tablename}{Table name, in format 'ProjectName.TableName',or
'TableName' (using current project).}

\item{partition}{Partition spec}
}
\description{
Get the size of table, unit is Byte.
}
\examples{
## get the size of 'sales'
\dontrun{rodps.table.size('sales')}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
  \code{\link{rodps.table.list}}, \code{\link{rodps.table.rows}},
  \code{\link{rodps.table.read}}, \code{\link{rodps.table.write}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
