% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_table.R
\name{rodps.table.write}
\alias{rodps.table.write}
\alias{rodps.write.table}
\title{Write Table}
\usage{
rodps.table.write(
  dataframe,
  full.tablename,
  partition = NULL,
  tablecomment = NULL,
  isdebug = FALSE,
  thread = 8
)

rodps.write.table(
  dataframe,
  full.tablename,
  partition = NULL,
  tablecomment = NULL,
  isdebug = FALSE,
  thread = 8
)
}
\arguments{
\item{dataframe}{Data in data.frame type, make sure the ColumnName is
allowable in ODPS.}

\item{full.tablename}{Table name, in format 'ProjectName.TableName',or
'TableName' (using current project).}

\item{partition}{Partition spec.}

\item{tablecomment}{Table comment.}

\item{isdebug}{Boolean value, if debugging is enabled.}

\item{thread}{Thread number.}
}
\description{
Write 'dataframe' into 'full.tablename' of ODPS, make sure the target table
'full.tablename' is not exist. Dataframe can be written to a non-exist table
or partition.
}
\examples{
## write data.frame into 'mytable'
\dontrun{ x<-data.frame(c1=1:10,c2=1:10)}
\dontrun{ rodps.table.write(x,'mytable')}
}
\seealso{
\code{\link{rodps.table.desc}}, \code{\link{rodps.table.drop}},
  \code{\link{rodps.table.exist}}, \code{\link{rodps.table.partitions}},
  \code{\link{rodps.table.list}}, \code{\link{rodps.table.rows}},
  \code{\link{rodps.table.size}}, \code{\link{rodps.table.read}}
}
\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
