% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_base_function.R
\name{rodps.unset}
\alias{rodps.unset}
\title{Unset task properties}
\usage{
rodps.unset(key)
}
\arguments{
\item{key}{setting name, e.g. odps.sql.allow.fullscan.}
}
\description{
Unset properties for SQL task
}
\examples{
## set full table scan to its default value
\dontrun{rodps.unset('odps.sql.allow.fullscan')}
}
\seealso{
\code{\link{RODPS}}, \code{\link{rodps.sql}},
  \code{\link{rodps.set}},
}
\author{
\email{ruibo.lirb@alibaba-inc.com}
}
