% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodps_misc.R
\name{rodps.vector}
\alias{rodps.vector}
\title{Set up odps.vector class, a vector is nothing but a column in table}
\usage{
rodps.vector(tblname, colname)
}
\description{
Set up odps.vector class, a vector is nothing but a column in table
}
