\name{rodps.project.current}
\alias{RODPS.PROJECT.CURRENT}
\alias{rodps.project.current}
\title{Current Project Name}
\usage{
rodps.project.current()
}
\description{
 get current project name.
}

\author{
\email{yunyuan.zhangyy@alibaba-inc.com}
}
\examples{
## get current project name
\dontrun{rodps.project.current()}
}
\seealso{
 \code{\link{rodps.project.use}}
}
\keyword{programming}
