% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_advanced_function_summary.R
\name{summary.rodps.data}
\alias{summary.rodps.data}
\title{store unique result in a temp table}
\usage{
\method{summary}{rodps.data}(rd)
}
\description{
store unique result in a temp table
}
