% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_advanced_function_head.R
\name{head.rodps.data}
\alias{head.rodps.data}
\title{Create odps.data and odps.vector in S4
store head result in a temp table}
\usage{
head.rodps.data(rd, n = 6L)
}
\description{
Create odps.data and odps.vector in S4
store head result in a temp table
}
