% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odps_advanced_function_hist.R
\name{hist.rodps.vector}
\alias{hist.rodps.vector}
\alias{rodps.table.hist}
\alias{rodps.hist}
\title{Extend hist hist returns a list of breaks, counts, density, mids, xname,
equidist, class attr generator a sql select count( case when ... end) b1,
count(case when ...) ... from tbl; then plot with the list}
\usage{
\method{hist}{rodps.vector}(ov)

rodps.table.hist(
  tablename,
  colname,
  breaks = NULL,
  freq = TRUE,
  include.lowest = TRUE,
  right = TRUE,
  main = paste("Histogram of ", colname),
  xlab = colname,
  ...
)

rodps.hist(
  tablename,
  colname,
  breaks = NULL,
  freq = TRUE,
  include.lowest = TRUE,
  right = TRUE,
  main = paste("Histogram of ", colname),
  xlab = colname,
  ...
)
}
\description{
Extend hist hist returns a list of breaks, counts, density, mids, xname,
equidist, class attr generator a sql select count( case when ... end) b1,
count(case when ...) ... from tbl; then plot with the list
}
